@@ -1,37 +1,32 @@
 package com.adaptris.filesystem;
 
-import static org.junit.Assert.assertFalse;
-import static org.junit.Assert.assertTrue;
+import static org.junit.jupiter.api.Assertions.assertFalse;
+import static org.junit.jupiter.api.Assertions.assertTrue;
 import java.io.File;
 import java.io.FileOutputStream;
 import java.io.IOException;
 import org.apache.commons.io.FileUtils;
-import org.junit.After;
-import org.junit.Before;
-import org.junit.Test;
+import org.junit.jupiter.api.AfterEach;
+import org.junit.jupiter.api.BeforeEach;
+import org.junit.jupiter.api.Test;
 import com.adaptris.core.AdaptrisMessage;
 import com.adaptris.core.AdaptrisMessageFactory;
-import com.adaptris.core.ServiceCase;
 import com.adaptris.core.common.ConstantDataInputParameter;
+import com.adaptris.interlok.junit.scaffolding.services.ExampleServiceCase;
 
 /**
  * @author mwarman
  */
-public class DeleteFileServiceTest extends ServiceCase {
+public class DeleteFileServiceTest extends ExampleServiceCase {
 
   private File directoryPath;
 
-  @Override
-  public boolean isAnnotatedForJunit4() {
-    return true;
-  }
-
-  @Before
+  @BeforeEach
   public void setUp() throws Exception {
     directoryPath = createTempDirectory();
   }
 
-  @After
+  @AfterEach
   public void tearDown() throws Exception {
     cleanUpTempDirectory(directoryPath);
   }
