@@ -1,26 +1,25 @@
 package com.adaptris.filesystem;
 
 import static org.hamcrest.core.IsInstanceOf.instanceOf;
-import static org.junit.Assert.assertEquals;
-import static org.junit.Assert.assertNotEquals;
-import static org.junit.Assert.assertNotSame;
-import static org.junit.Assert.assertThat;
-import static org.junit.Assert.assertTrue;
+import static org.junit.jupiter.api.Assertions.assertEquals;
+import static org.junit.jupiter.api.Assertions.assertNotEquals;
+import static org.junit.jupiter.api.Assertions.assertNotSame;
+import static org.junit.jupiter.api.Assertions.assertTrue;
 
 import java.io.File;
 import java.io.IOException;
 import java.util.Date;
 import java.util.GregorianCalendar;
 
-import org.junit.After;
-import org.junit.Before;
-import org.junit.Test;
+import org.junit.jupiter.api.AfterEach;
+import org.junit.jupiter.api.BeforeEach;
+import org.junit.jupiter.api.Test;
 
 public class DirectoryEntityTest {
 
     private File tempDir;
 
-    @Before
+    @BeforeEach
     public void setUp() throws IOException {
         tempDir = File.createTempFile(DirectoryEntityTest.class.getSimpleName(), "", null);
         tempDir.delete();
@@ -29,7 +28,7 @@ public class DirectoryEntityTest {
         }
     }
 
-    @After
+    @AfterEach
     public void tearDown() {
         for (final File f : tempDir.listFiles()) {
             f.delete();
@@ -56,11 +55,11 @@ public class DirectoryEntityTest {
         Date newDate = gCal.getTime();
         directoryEntity.setCreatedAt(newDate);
         //Confirming that directoryEntity is set to defaults
-        assertEquals("Should have same \'getDescription\'", directoryEntity.getDescription(), directoryEntityII.getDescription());
-        assertEquals("Should have identical \'updatedAt\' dates", directoryEntity.getUpdatedAt(), directoryEntityII.getUpdatedAt());
-        assertNotSame("Should have disparate \'getId\' prior to update", directoryEntity.getId(), directoryEntityII.getId());
-        assertNotSame("Should have disparate \'getCreatedAt\'", directoryEntity.getCreatedAt(), directoryEntityII.getCreatedAt());
-        assertNotSame("Should be two unique instances of \'DirectoryEntity\'", directoryEntity, directoryEntityII);
+        assertEquals(directoryEntity.getDescription(), directoryEntityII.getDescription(), "Should have same \'getDescription\'");
+        assertEquals(directoryEntity.getUpdatedAt(), directoryEntityII.getUpdatedAt(), "Should have identical \'updatedAt\' dates");
+        assertNotSame(directoryEntity.getId(), directoryEntityII.getId(), "Should have disparate \'getId\' prior to update");
+        assertNotSame(directoryEntity.getCreatedAt(), directoryEntityII.getCreatedAt(), "Should have disparate \'getCreatedAt\'");
+        assertNotSame(directoryEntity, directoryEntityII, "Should be two unique instances of \'DirectoryEntity\'");
         //Setting new attributes
             directoryEntity.setId("14");
             directoryEntityII.setId("14");
@@ -70,13 +69,13 @@ public class DirectoryEntityTest {
             directoryEntity.setUpdatedAt(newDate);
             directoryEntity.setSize(5000L);
         //Confirming that directoryEntity is updated from defaults
-        assertEquals("Should return \'Hello World\' as \'getParentDirectory\'","Hello World", directoryEntity.getParentDirectory());
-        assertEquals("Should return matching \'getId\'", directoryEntity.getId(), directoryEntityII.getId());
-        assertNotEquals("\'UpdatedAt\' should no longer match", directoryEntity.getUpdatedAt(), directoryEntityII.getUpdatedAt());
-        assertNotEquals("Should have disparate \'getDescription\' after update" , directoryEntity.getDescription(), directoryEntityII.getDescription());
-        assertNotEquals("Should have disparate \'getSize\'", directoryEntity.getSize(), directoryEntityII.getSize());
-        assertTrue("Should match \'newDate\' as the \'getCreatedAt\'", directoryEntity.toString().contains("\"createdAt\":\"2018-12-23T12:30:00+0000\""));
-        assertTrue("Should still be a string type", directoryEntity.toJSON().contains("\"createdAt\":\"2018-12-23T12:30:00+0000\""));
-        assertThat("\'toString\' should return String type", directoryEntity.toJSON(), instanceOf(String.class));
+        assertEquals("Hello World", directoryEntity.getParentDirectory(), "Should return \'Hello World\' as \'getParentDirectory\'");
+        assertEquals(directoryEntity.getId(), directoryEntityII.getId(), "Should return matching \'getId\'");
+        assertNotEquals(directoryEntity.getUpdatedAt(), directoryEntityII.getUpdatedAt(), "\'UpdatedAt\' should no longer match");
+        assertNotEquals(directoryEntity.getDescription(), directoryEntityII.getDescription(), "Should have disparate \'getDescription\' after update");
+        assertNotEquals(directoryEntity.getSize(), directoryEntityII.getSize(), "Should have disparate \'getSize\'");
+        assertTrue(directoryEntity.toString().contains("\"createdAt\":\"2018-12-23T12:30:00+0000\""), "Should match \'newDate\' as the \'getCreatedAt\'");
+        assertTrue(directoryEntity.toJSON().contains("\"createdAt\":\"2018-12-23T12:30:00+0000\""), "Should still be a string type");
+        assertEquals(directoryEntity.toJSON().getClass(), String.class, "\'toString\' should return String type");
     }
 }
